requirements.txt
nose2
assertpy~=1.1
pytest~=6.2.5
setuptools~=49.2.1
pyhamcrest
coverage
parameterized~=0.7.4
